// .tmp/zip_snapshot/tote-bag-practicantes/apps/api/src/generated/client/enums.ts
export const TransactionType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionCategory = {
  SALE: 'SALE',
  PURCHASE: 'PURCHASE',
  OPEX: 'OPEX',
  PAYROLL: 'PAYROLL'
} as const

export type TransactionCategory = (typeof TransactionCategory)[keyof typeof TransactionCategory]